+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Initial boilerplate.
 - Add git parameter to localize refs get
 - Add commands to export
+- Add publish\_to: none when localizing
 
 ### Changed